Move FilterListbox component to shared UI
