+++ apps/ff-site/src/app/_components/CategoryFilter.tsx
@@ -1,13 +1,13 @@
 'use client'
 
 import { DEFAULT_CATEGORY_FILTER_OPTION } from '@filecoin-foundation/hooks/useFilter'
+import { FilterListbox } from '@filecoin-foundation/ui/FilterListbox'
 import type { OptionType } from '@filecoin-foundation/ui/Listbox/ListboxOption'
 import { CATEGORY_KEY } from '@filecoin-foundation/utils/constants/urlParamsConstants'
 
 import { useListboxQueryState } from '@/hooks/useListboxQueryState'
 
 import { CategorySidebar } from '@/components/CategorySidebar'
-import { FilterListbox } from '@/components/FilterListbox'
 
 type CategoryProps = {
   options: Array<OptionType>
