+++ apps/ff-site/src/app/_components/LocationFilter.tsx
@@ -1,13 +1,12 @@
 'use client'
 
 import { DEFAULT_LOCATION_FILTER_OPTION } from '@filecoin-foundation/hooks/useFilter'
+import { FilterListbox } from '@filecoin-foundation/ui/FilterListbox'
 import type { OptionType } from '@filecoin-foundation/ui/Listbox/ListboxOption'
 import { LOCATION_KEY } from '@filecoin-foundation/utils/constants/urlParamsConstants'
 import { GlobeSimple } from '@phosphor-icons/react'
 
 import { useListboxQueryState } from '@/hooks/useListboxQueryState'
-
-import { FilterListbox } from '@/components/FilterListbox'
 
 type LocationFilterProps = {
   options: Array<OptionType>
