+++ packages/ui/package.json
@@ -22,6 +22,7 @@
     "./DescriptionText": "./src/DescriptionText.tsx",
     "./DigestArticleHeader": "./src/DigestArticleHeader.tsx",
     "./ErrorMessage": "./src/ErrorMessage.tsx",
+    "./FilterListbox": "./src/FilterListbox.tsx",
     "./GlobalError": "./src/GlobalError.tsx",
     "./Heading": "./src/Heading.tsx",
     "./Icon": "./src/Icon.tsx",
